% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment.R
\name{as_SummarizedExperiment}
\alias{as_SummarizedExperiment}
\alias{as_SummarizedExperiment,SumExp-method}
\title{Coerce a SumExp object to a SummarizedExperiment object}
\usage{
as_SummarizedExperiment(x, ...)

\S4method{as_SummarizedExperiment}{SumExp}(x, ...)
}
\arguments{
\item{x}{A \code{\link{SumExp}} object}

\item{...}{Not used}
}
\value{
A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object with the row and column data
}
\description{
Coerce a SumExp object to a SummarizedExperiment object
}
