% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{as_tibble,SumExp-method}
\alias{as_tibble,SumExp-method}
\title{Coerce a SumExp object to a `tibble`}
\usage{
\S4method{as_tibble}{SumExp}(x, ...)
}
\arguments{
\item{x}{A `SumExp` object}
}
\value{
A `tibble` with the row and column data frames as well as the matrices. Each matrix
  in the `SumExp` object is represented as a long format tibble with columns `.row_id`,
  `.col_id`, and the matrix name.
}
\description{
Coerce a SumExp object to a `tibble`
}
