% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble,SumExp-method}
\title{Coerce a SumExp object to a \code{tibble}}
\usage{
as_tibble(x, ...)

\S4method{as_tibble}{SumExp}(x, ...)
}
\arguments{
\item{x}{A \code{\link{SumExp}} object}

\item{...}{Not used}
}
\value{
A \code{tibble} with the row and column data frames as well as the matrices. Each matrix
in the \code{\link{SumExp}} object is represented as a long format tibble with columns \code{.row_id},
\code{.col_id}, and the matrix name.
}
\description{
Coerce a SumExp object to a \code{tibble}
}
