% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{copy_labels_list}
\alias{copy_labels_list}
\title{Copy labels handling nested lists}
\usage{
copy_labels_list(from, to)
}
\arguments{
\item{from}{A vector or a data.frame (or tibble) to copy labels from.}

\item{to}{A vector or data.frame (or tibble) to copy labels to.}
}
\description{
\code{\link[labelled:copy_labels]{labelled::copy_labels()}} does not handle nested lists. This function is a workaround.
}
