% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumExp-class.R
\name{.print_data_frame_like_tibble}
\alias{.print_data_frame_like_tibble}
\title{Print a data frame like printing a tibble}
\usage{
.print_data_frame_like_tibble(x, ..., n = 5)
}
\arguments{
\item{x}{A data frame}

\item{..., n}{Additional arguments passed to \code{\link[tibble:formatting]{tibble::print.tbl()}}}
}
\value{
The input data frame
}
\description{
Print a data frame like printing a tibble
}
\examples{
.print_data_frame_like_tibble(iris)
}
