% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmpl_se}
\alias{exmpl_se}
\title{An example dataset for SumExp class}
\format{
A \code{SumExp} object
}
\usage{
exmpl_se
}
\description{
An example dataset for SumExp class
}
\keyword{datasets}
