% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumExp-class.R
\name{Extract}
\alias{Extract}
\title{Extract elements from a `SumExp` object}
\usage{
\S4method{[}{SumExp}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A `SumExp` object}

\item{i, j}{Indices specifying elements to extract. The indices can be numeric, character or
logical vectors. The expression `i` is evaluated in the context of `x@row_df` and `j` in
the context of `x@col_df`.}
}
\description{
Extract elements from a `SumExp` object
}
