% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{ggplot,SumExp-method}
\alias{ggplot,SumExp-method}
\title{Create a ggplot object from a SumExp object}
\usage{
\S4method{ggplot}{SumExp}(data, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{A `SumExp` object}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Used prior to tidy
evaluation.}
}
\value{
A ggplot object
}
\description{
Create a ggplot object from a SumExp object
}
