% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListMatrix-class.R
\docType{class}
\name{ListMatrix-class}
\alias{ListMatrix-class}
\alias{ListMatrix}
\alias{dim,ListMatrix-method}
\alias{nrow}
\alias{nrow,ListMatrix-method}
\alias{ncol}
\alias{ncol,ListMatrix-method}
\alias{dimnames,ListMatrix-method}
\alias{dimnames<-,ListMatrix-method}
\alias{colnames,ListMatrix-method}
\alias{colnames<-,ListMatrix-method}
\alias{rownames,ListMatrix-method}
\alias{rownames<-,ListMatrix-method}
\alias{show,ListMatrix-method}
\alias{print,ListMatrix-method}
\alias{names,ListMatrix-method}
\alias{[[,ListMatrix-method}
\alias{[[<-,ListMatrix-method}
\title{A class for storing a list of matrices}
\usage{
\S4method{dim}{ListMatrix}(x)

nrow(x)

\S4method{nrow}{ListMatrix}(x)

ncol(x)

\S4method{ncol}{ListMatrix}(x)

\S4method{dimnames}{ListMatrix}(x)

\S4method{dimnames}{ListMatrix}(x) <- value

\S4method{colnames}{ListMatrix}(x)

\S4method{colnames}{ListMatrix}(x) <- value

\S4method{rownames}{ListMatrix}(x)

\S4method{rownames}{ListMatrix}(x) <- value

\S4method{show}{ListMatrix}(object)

\S4method{print}{ListMatrix}(x)

\S4method{names}{ListMatrix}(x)

\S4method{[[}{ListMatrix}(x, i)

\S4method{[[}{ListMatrix}(x, i) <- value
}
\arguments{
\item{x}{A `ListMatrix` object}
}
\description{
A class for storing a list of matrices
}
\section{Slots}{

\describe{
\item{\code{matrices}}{A list with the matrices}
}}

