% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListMatrix-class.R
\name{name_labs,ListMatrix-method}
\alias{name_labs,ListMatrix-method}
\title{Get the names of the matrices in a `ListMatrix` with label attributes}
\usage{
\S4method{name_labs}{ListMatrix}(x)
}
\arguments{
\item{x}{A `ListMatrix` object}
}
\value{
A character vector with the names of the matrices. If the matrices have label
  attributes, the labels are returned as the names of the returned character vector.
}
\description{
Get the names of the matrices in a `ListMatrix` with label attributes
}
