% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-method.R
\name{name_labs}
\alias{name_labs}
\title{Get names and label attributes of a list}
\usage{
name_labs(x)
}
\arguments{
\item{x}{A list with names and optionally label attributes}
}
\value{
A character vector with the names. If elements have label attributes by a method of
\code{\link{labelled}} package, the labels are returned as the names of the returned character
vector.
}
\description{
Get names and label attributes of a list
}
\examples{
  x <- list(a = 1:3, b = 4:6)
  labelled::var_label(x$a) <- "A"
  name_labs(x)
}
