% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_columns}
\alias{split_columns}
\alias{split_rows}
\title{Divide into groups}
\usage{
split_columns(x, f)

split_rows(x, f)
}
\arguments{
\item{x}{A matrix-like object, e.g. \code{\link{SumExp}}}

\item{f}{A factor vector to split by}
}
\value{
A list of matrix-like objects
}
\description{
Split a matrix-like object, e.g. \code{\link{SumExp}}, into groups based on a factor
column-/row-wise.
}
\examples{
se <- SumExp::exmpl_se
split_rows(se, row_df(se)$color)
split_columns(se, rep(1:2, each = ncol(se) / 2))
split_columns(exmpl_se, f = col_df(exmpl_se)$type)
}
