% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{[,SumExp-method}
\alias{[,SumExp-method}
\title{Extract a part from a \code{\link{SumExp}} object}
\usage{
\S4method{[}{SumExp}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{\link{SumExp}} object}

\item{i, j}{Indices specifying elements to extract. The indices can be numeric, character or
logical vectors. If given as a quoted expression (ie. \code{\link{call}} or \code{\link{quosure}} object), it
will be evaluated in the context of the \code{\link{SumExp}} object. The \code{i} is evaluated in the
context of \code{row_df(x)} and \code{j} in the context of \code{col_df(x)}.}

\item{...}{Not used}

\item{drop}{Not supported. Always \code{FALSE}}
}
\description{
Extract a part from a \code{\link{SumExp}} object
}
