% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{subset_copy_labels.data.frame}
\alias{subset_copy_labels.data.frame}
\title{Subset and copy labels handling nested lists}
\usage{
subset_copy_labels.data.frame(from, idx)
}
\arguments{
\item{from}{A data frame}

\item{idx}{Index to subset \code{from}}
}
\description{
Subset and copy labels handling nested lists
}
