% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumExp-class.R
\docType{class}
\name{SumExp-class}
\alias{SumExp-class}
\alias{SumExp}
\alias{rowData}
\alias{rowData,SumExp-method}
\alias{rowData<-}
\alias{rowData<-,SumExp,data.frame-method}
\alias{colData}
\alias{colData,SumExp-method}
\alias{colData<-}
\alias{colData<-,SumExp,data.frame-method}
\alias{metadata}
\alias{metadata,SumExp-method}
\alias{metadata<-}
\alias{metadata<-,SumExp,list-method}
\alias{assay}
\alias{assay,SumExp-method}
\alias{assay<-}
\alias{assay<-,SumExp,ANY,matrix-method}
\alias{[,SumExp-method}
\alias{as_tibble}
\alias{as_tibble,SumExp-method}
\alias{ggplot}
\alias{ggplot,SumExp-method}
\title{An light S4 class for the matrices with additional information for columns and rows}
\usage{
rowData(x)

\S4method{rowData}{SumExp}(x)

rowData(x) <- value

\S4method{rowData}{SumExp,data.frame}(x) <- value

colData(x)

\S4method{colData}{SumExp}(x)

colData(x) <- value

\S4method{colData}{SumExp,data.frame}(x) <- value

metadata(x)

\S4method{metadata}{SumExp}(x)

metadata(x) <- value

\S4method{metadata}{SumExp,list}(x) <- value

assay(x, i, ...)

\S4method{assay}{SumExp}(x, i, ...)

assay(x, i, ...) <- value

\S4method{assay}{SumExp,ANY,matrix}(x, i, ...) <- value

\S4method{[}{SumExp}(x, i, j, ..., drop = FALSE)

as_tibble(x, ...)

\S4method{as_tibble}{SumExp}(x, ...)

ggplot(data, mapping = aes(), ..., environment = parent.frame())

\S4method{ggplot}{SumExp}(data, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{x}{A `SumExp` object}
}
\description{
An light S4 class for the matrices with additional information for columns and rows

Methods for `SumExp` objects
}
\section{Slots}{

\describe{
\item{\code{matrices}}{A `ListMatrix` object, a list with the matrices}

\item{\code{colData}}{A data frame with additional information for columns}

\item{\code{rowData}}{A data frame with additional information for rows}

\item{\code{metadata}}{A `list` with any other additional information}
}}

