% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListMatrix-class.R
\name{Extract}
\alias{Extract}
\alias{[[,ListMatrix-method}
\alias{[[<-,ListMatrix-method}
\title{Extract or Replace an element of a `ListMatrix`}
\usage{
\S4method{[[}{ListMatrix}(x, i, j, ...)

\S4method{[[}{ListMatrix}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A `ListMatrix` object}

\item{i}{An index or a name of the element to extract/replace}
}
\description{
Extract or Replace an element of a `ListMatrix`
}
