% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListMatrix-class.R
\docType{class}
\name{ListMatrix-class}
\alias{ListMatrix-class}
\alias{ListMatrix}
\alias{names,ListMatrix-method}
\alias{as.list,ListMatrix-method}
\alias{[[<-,ListMatrix,ANY,ANY-method}
\alias{[[<-,ListMatrix-method}
\alias{dim,ListMatrix-method}
\alias{nrow,ListMatrix-method}
\alias{ncol,ListMatrix-method}
\alias{dimnames,ListMatrix-method}
\alias{dimnames<-,ListMatrix,ANY-method}
\alias{dimnames<-,ListMatrix-method}
\alias{colnames,ListMatrix-method}
\alias{colnames<-,ListMatrix-method}
\alias{rownames,ListMatrix-method}
\alias{rownames<-,ListMatrix-method}
\alias{show,ListMatrix-method}
\alias{print,ListMatrix-method}
\alias{t,ListMatrix-method}
\title{A class for storing a list of matrices}
\usage{
\S4method{names}{ListMatrix}(x)

\S4method{as.list}{ListMatrix}(x)

\S4method{[[}{ListMatrix,ANY,ANY}(x, i) <- value

\S4method{dim}{ListMatrix}(x)

\S4method{nrow}{ListMatrix}(x)

\S4method{ncol}{ListMatrix}(x)

\S4method{dimnames}{ListMatrix}(x)

\S4method{dimnames}{ListMatrix,ANY}(x) <- value

\S4method{colnames}{ListMatrix}(x)

\S4method{colnames}{ListMatrix}(x) <- value

\S4method{rownames}{ListMatrix}(x)

\S4method{rownames}{ListMatrix}(x) <- value

\S4method{show}{ListMatrix}(object)

\S4method{print}{ListMatrix}(x)

\S4method{t}{ListMatrix}(x)
}
\arguments{
\item{x, object}{A \code{\link{ListMatrix}} object}

\item{i}{An integer or a character}

\item{value}{A matrix for \verb{[[<-};
A character vector for \verb{colnames<-} and \verb{rownames<-};
A list of row and column names for \verb{dimnames<-}}
}
\description{
This class inherits \code{\link[base:list]{base::list}} class and is used to store a list of matrices. This class
makes sure that all matrices are named and have the same dimensions and row/column names.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list with matrices. All matrices must have the same dimensions and row/column
names.}

\item{\code{names}}{A character vector with the names of the matrices. This is added because the
\code{.Data} doesn't contain the names.}
}}

\examples{
nms <- list(LETTERS[1:4], letters[1:5])
m1 <- matrix(sample(20), nrow = 4, dimnames = nms)
m2 <- matrix(sample(LETTERS, 20), nrow = 4, dimnames = nms)
l <- ListMatrix('mat1' = m1, 'mat2' = m2)

# Get the names
names(l)
# Get as list
x <- as.list(l)
str(x)
# Update a matrix in a ListMatrix object
m3 <- m2
m3[2, ] <- c(letters[5:9])
l[[2]] <- m3
l
# Get the dimensions of a ListMatrix object
dim(l)
nrow(l)
ncol(l)
}
