% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumExp-class.R
\docType{class}
\name{SumExp-class}
\alias{SumExp-class}
\alias{SumExp}
\alias{row_df}
\alias{row_df,SumExp-method}
\alias{row_df<-}
\alias{row_df<-,SumExp,data.frame-method}
\alias{col_df}
\alias{col_df,SumExp-method}
\alias{col_df<-}
\alias{col_df<-,SumExp,data.frame-method}
\alias{metadata}
\alias{metadata,SumExp-method}
\alias{metadata<-}
\alias{metadata<-,SumExp,list-method}
\title{An light S4 class for the matrices with additional information for columns and rows}
\usage{
row_df(x)

\S4method{row_df}{SumExp}(x)

row_df(x) <- value

\S4method{row_df}{SumExp,data.frame}(x) <- value

col_df(x)

\S4method{col_df}{SumExp}(x)

col_df(x) <- value

\S4method{col_df}{SumExp,data.frame}(x) <- value

metadata(x)

\S4method{metadata}{SumExp}(x)

metadata(x) <- value

\S4method{metadata}{SumExp,list}(x) <- value
}
\arguments{
\item{x}{A `SumExp` object}
}
\description{
An light S4 class for the matrices with additional information for columns and rows

Methods for `SumExp` objects
}
\section{Slots}{

\describe{
\item{\code{matrices}}{A list with matrices. All matrices must have the same dimensions and
row/column names.}

\item{\code{col_df}}{A data frame with additional information for columns}

\item{\code{row_df}}{A data frame with additional information for rows}

\item{\code{metadata}}{A `list` with any other additional information}
}}

